pyproject.toml
[tool.poetry]
name = "raspi-test"
version = "0.1.0"
description = ""
authors = ["cykps <cykps.ps@gmail.com>"]
readme = "README.md"
packages = [{include = "raspi_test"}]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
pyaudio = "^0.2.14"
rpi-gpio = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
